Refactor Registration module and update Starter app
Removed the WeatherForecast file from the Registration module. User model was renamed and updated in the Registration namespace. Several files in both the Registration and Starter directories were updated. New files (InMemoryFakes.fs, FacadesCreator.fs, Services.fs) were added to the Starter directory to help with the creation of facades. The launch settings for the Starter application were also updated.